+++ source/AutoSizer/AutoSizer.test.js
@@ -12,7 +12,7 @@ describe('AutoSizer', () => {
   beforeAll(() => jasmine.clock().install())
   afterAll(() => jasmine.clock().uninstall())
 
-  // Used by the renderOrUpdateList() helper method
+  // Used by the renderOrUpdateComponent() helper method
   // Unless we attach the node to body, getBoundingClientRect() won't work
   var node = null
   beforeEach(() => {
@@ -27,8 +27,10 @@ describe('AutoSizer', () => {
     useReactChild,
     {
       bar = 123,
+      className = undefined,
       foo = 456,
       height = 100,
+      styleSheet = undefined,
       width = 200
     } = {}
   ) {
@@ -36,7 +38,10 @@ describe('AutoSizer', () => {
 
     if (useReactChild) {
       autoSizer = (
-        <AutoSizer>
+        <AutoSizer
+          className={className}
+          styleSheet={styleSheet}
+        >
           <ChildComponent
             bar={bar}
             foo={foo}
@@ -48,7 +53,9 @@ describe('AutoSizer', () => {
         <AutoSizer
           ChildComponent={ChildComponent}
           bar={bar}
+          className={className}
           foo={foo}
+          styleSheet={styleSheet}
         />
       )
     }
@@ -60,15 +67,15 @@ describe('AutoSizer', () => {
     )
   }
 
-  function renderOrUpdateList (useReactChild, props) {
-    let rendered = render(getMarkup(props), node)
+  function renderOrUpdateComponent (useReactChild, props) {
+    let rendered = render(getMarkup(useReactChild, props), node)
 
     return findDOMNode(rendered)
   }
 
   it('should relay properties to ChildComponent or React child', () => {
     [false, true].forEach(useReactChild => {
-      const component = renderOrUpdateList(useReactChild)
+      const component = renderOrUpdateComponent(useReactChild)
       expect(findDOMNode(component).textContent).toContain('foo:456')
       expect(findDOMNode(component).textContent).toContain('bar:123')
     })
@@ -76,10 +83,39 @@ describe('AutoSizer', () => {
 
   it('should set the correct initial width and height of ChildComponent or React child', () => {
     [false, true].forEach(useReactChild => {
-      const component = renderOrUpdateList(useReactChild)
+      const component = renderOrUpdateComponent(useReactChild)
       let domNode = findDOMNode(component)
       expect(domNode.textContent).toContain('height:100')
       expect(domNode.textContent).toContain('width:200')
+    })
+  })
+
+  describe('styles and classeNames', () => {
+    it('should use the expected global CSS classNames', () => {
+      const node = renderOrUpdateComponent()
+      expect(node.querySelector('.AutoSizer')).toBeTruthy()
+    })
+
+    it('should use a custom :className if specified', () => {
+      const node = renderOrUpdateComponent(true, { className: 'foo' })
+      expect(node.querySelector('.AutoSizer').className).toContain('foo')
+    })
+
+    it('should use custom :styleSheet if specified', () => {
+      const node = renderOrUpdateComponent(true, {
+        styleSheet: {
+          AutoSizer: { color: 'red' }
+        }
+      })
+      expect(node.querySelector('.AutoSizer').style.color).toEqual('red')
+    })
+
+    it('should use overriden static styles', () => {
+      const backup = { ...AutoSizer.defaultStyleSheet }
+      AutoSizer.defaultStyleSheet.AutoSizer = { color: 'blue' }
+      const node = renderOrUpdateComponent()
+      expect(node.querySelector('.AutoSizer').style.color).toEqual('blue')
+      AutoSizer.defaultStyleSheet = backup
     })
   })