+++ source/utils.js
@@ -1,3 +1,7 @@
+import Prefixer from 'inline-style-prefixer'
+
+const prefixer = new Prefixer()
+
 /**
  * Binary search function inspired by react-infinite.
  */
@@ -169,4 +173,22 @@ export function initOnRowsRenderedHelper () {
       onRowsRendered({ startIndex, stopIndex })
     }
   }
+}
+
+/**
+ * Adds vender prefixes to a style object.
+ */
+export function prefixStyle (style) {
+  return prefixer.prefix(style)
+}
+
+/**
+ * Adds vender prefixes for all of the styles in a stylesheet and returns a prefixed copy.
+ */
+export function prefixStyleSheet (styleSheet) {
+  const prefixedStyleSheet = {}
+  for (var style in styleSheet) {
+    prefixedStyleSheet[style] = prefixStyle(styleSheet[style])
+  }
+  return prefixedStyleSheet
 }